requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
hazelcast-python-client==5.3.0
idna==3.4
pydantic==2.4.2
pydantic-core==2.10.1
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.23.2
